+++ src/Symfony/Bundle/SecurityBundle/Tests/DependencyInjection/Fixtures/php/merge_import.php
@@ -6,7 +6,6 @@ $container->loadFromExtension('security', array(
             'form_login' => array(
                 'login_path' => '/login',
             ),
-        'logout_on_user_change' => true,
         ),
     ),
     'role_hierarchy' => array(