+++ src/Symfony/Component/Translation/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+4.1.0
+-----
+
+ * The `FileDumper::setBackup()` method is deprecated and will be removed in 5.0.
+ * The `TranslationWriter::disableBackup()` method is deprecated and will be removed in 5.0.
+
 4.0.0
 -----